requirements.txt
python-dateutil
flask
flask_wtf
bootstrap-flask
wtforms
waitress
pony
apscheduler
psycopg2
requests
urllib3<2
oauthlib
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib